requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
defusedxml==0.7.1
distlib==0.3.9
Django==5.1.3
django-allauth==65.3.1
django-cors-headers==4.6.0
django-countries==7.6.1
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-filter==24.3
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
gunicorn==23.0.0
idna==3.10
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
python3-openid==3.2.0
requests==2.32.3
sqlparse==0.5.2
stripe==11.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
virtualenv==20.28.0
whitenoise==6.8.2
